+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - add support for policy object security url allow list
 - add support for policy object security Advanced Malware Protection Profile
 - add support for service OSPFv3 IPv6
+- add support for service Multicast
 - add support for `cloud_qos` and `cloud_qos_service_side` in `sdwan_localized_policy`
 - add `topology_label` attribute to configuration group
 - fix service bgp and transport bgp failing when ipv6 neighbor has `maximum_prefix_reach_policy` set to `off`